# docs/api/swagger.yaml
openapi: 3.1.0
info:
  title: Distributed Task Manager API
  description: API specification for a cloud-native, serverless task management platform featuring distributed microservices and streaming video/audio attachments.
  version: 1.1.0
  contact:
    name: API Support
    email: support@taskmanager.local
servers:
  - url: https://api.taskmanager.local/v1
    description: Production API Gateway
  - url: https://staging.api.taskmanager.local/v1
    description: Staging Environment
  - url: http://localhost:3000/v1
    description: Local Development Environment
...
paths:
...
  /auth/password/recovery:
    post:
      tags:
        - Authentication
      summary: Request password recovery email
      operationId: requestPasswordRecovery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '202':
          description: Recovery email dispatched (if email exists)
  /auth/password/reset:
    post:
      tags:
        - Authentication
      summary: Reset password using recovery token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token:
                  type: string
                newPassword:
                  type: string
                  format: password
                  minLength: 8
      responses:
        '200':
          description: Password reset successful
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /tasks/{id}/assignments:
    get:
      tags:
        - Assignments
      summary: List users assigned to a task
      operationId: getTaskAssignments
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
      responses:
        '200':
          description: List of assigned users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags:
        - Assignments
      summary: Assign a user to a task
      operationId: assignUserToTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: string
                  format: uuid
      responses:
        '201':
          description: User assigned
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /tasks/{id}/assignments/{userId}:
    delete:
      tags:
        - Assignments
      summary: Remove a user assignment from a task
      operationId: removeTaskAssignment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Assignment removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /tasks/{id}/comments:
    get:
      tags:
        - Collaboration
      summary: Get comments for a task
      operationId: getTaskComments
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageCursor'
      responses:
        '200':
          description: Paginated comments
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CursorPaginatedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Comment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags:
        - Collaboration
      summary: Add a comment to a task
      operationId: addTaskComment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentInput'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /notifications:
    get:
      tags:
        - Notifications
      summary: Get user notifications
      operationId: getNotifications
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageCursor'
        - name: isRead
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CursorPaginatedResponse'
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Notification'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /notifications/{id}/read:
    patch:
      tags:
        - Notifications
      summary: Mark a single notification as read
      operationId: markNotificationRead
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Notification marked as read
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /notifications/read-all:
    patch:
      tags:
        - Notifications
      summary: Mark all notifications as read
      operationId: markAllNotificationsRead
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All notifications marked as read
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /roles:
    get:
      tags:
        - Admin
      summary: List all available roles (Admin Only)
      operationId: listRoles
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{id}/roles:
    get:
      tags:
        - Admin
      summary: Get roles for a specific user (Admin Only)
      operationId: getUserRoles
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
    put:
      tags:
        - Admin
      summary: Update roles for a specific user (Admin Only)
      operationId: updateUserRoles
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roleIds]
              properties:
                roleIds:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: Roles updated successfully
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Authorization header using the Bearer scheme.
...
  parameters:
    TaskIdParam:
      name: id
      in: path
      required: true
      description: UUID of the task
      schema:
        type: string
        format: uuid
...
    PageLimit:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    PageCursor:
      name: cursor
      in: query
      description: Cursor for pagination to fetch the next set of results
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request payload format
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: User lacks sufficient permissions
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Conflict:
      description: Resource already exists or state conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    UnprocessableEntity:
      description: Validation error on the request content
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
    ProblemDetails:
      type: object
      description: RFC 9457 Problem Details for HTTP APIs
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          example: 'https://api.taskmanager.local/errors/validation-error'
        title:
          type: string
          example: 'Validation Error'
        status:
          type: integer
          example: 422
        detail:
          type: string
          example: 'The provided email address is invalid.'
        instance:
          type: string
          format: uri
          example: '/auth/register'
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    CursorPaginatedResponse:
      type: object
      required: [limit]
      properties:
        limit:
          type: integer
          example: 20
        nextCursor:
          type: string
          nullable: true
          description: Token to fetch the next page. Null if no more pages.
        total:
          type: integer
          description: Optional total count if requested or available
...
    Role:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 'ADMIN'
        description:
          type: string
    User:
      type: object
      required: [id, username, email, isActive, createdAt]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        fullName:
          type: string
        avatarUrl:
          type: string
          format: uri
        isActive:
          type: boolean
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        createdAt:
          type: string
          format: date-time
...
    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        parentCommentId:
          type: string
          format: uuid
        mentions:
          type: array
          items:
            type: string
            format: uuid
            description: Array of user IDs mentioned in the comment
    Comment:
      allOf:
        - $ref: '#/components/schemas/CommentInput'
        - type: object
          required: [id, taskId, userId, createdAt]
          properties:
            id:
              type: string
              format: uuid
            taskId:
              type: string
              format: uuid
            userId:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    Notification:
      type: object
      required: [id, type, title, message, isRead, createdAt]
      properties:
        id:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        commentId:
          type: string
          format: uuid
        type:
          type: string
          enum: [ASSIGNMENT, TASK_UPDATED, MENTION]
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
        readAt:
          type: string
          format: date-time
...